Refined single syntax
Simplified the single syntax by defining it as a pair rather than as the 1-tuple object. Meaning execution of a single in the eson document is now a first class operation instead of a special case. This means we now have two types of singles:- those that appear as pairs and those that appear as 1-tuple JSON objects. The 1-tuple object single is called a value single  because it results in substitution, pair singles are removed after evaluation.